// src/models/schema/Complaint.schema.d.ts
import { IComplaint } from "../../types/comlpaint.type";
export declare const ComplaintModel: import("mongoose").Model<IComplaint, {}, {}, {}, import("mongoose").Document<unknown, {}, IComplaint, {}, {}> & IComplaint & Required<{
    _id: import("mongoose").Types.ObjectId;
}> & {
    __v: number;
}, any>;
//# sourceMappingURL=Complaint.schema.d.ts.map